requirements.txt
asgiref==3.2.10
astroid==2.6.6
boto3==1.18.24
botocore==1.21.24
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.5
dj-database-url==0.5.0
Django==3.1
django-admin-thumbnails==0.2.6
django-decouple==2.1
django-heroku==0.3.1
django-honeypot==1.0.1
django-on-heroku==1.1.2
django-request==1.5.6
django-session-timeout==0.1.0
django-smtp-ssl==1.0
django-storages==1.11.1
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==3.2
isort==5.9.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==9.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.3
pylint==2.9.6
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
